+++ example/lib/example5_themes_playground/widgets/panels/component_settings/component_settings.dart
@@ -29,6 +29,22 @@ class ComponentSettings extends StatelessWidget {
         // UseMaterial3TextTheme(controller: controller),
         UseTinted3TextTheme(controller: controller),
         SwitchListTile(
+          title: const Text('Use Material 2 style Divider in Material 3'),
+          subtitle: const Text('In M3 used primary color tinted outlineVariant '
+              'does not fit on all colors. The M2 style based on '
+              'black or white with opacity does. It is also less prominent '
+              'than the M3 style and may be preferred.'),
+          value: controller.useM2StyleDividerInM3 &&
+              controller.useSubThemes &&
+              controller.useMaterial3 &&
+              controller.useFlexColorScheme,
+          onChanged: controller.useSubThemes &&
+                  controller.useFlexColorScheme &&
+                  controller.useMaterial3
+              ? controller.setUseM2StyleDividerInM3
+              : null,
+        ),
+        SwitchListTile(
           title: const Text('Themed interaction effects'),
           subtitle: const Text('The hover, focus, highlight, splash and '
               'disabled colors and effects are based on primary color.\n'